Add an option to handle items in bulk.
In some cases the provider might be able to perform better with items
in bulk rather than individual requests. For example, Elasticsearch
is more performant when indexing/updating multiple records in a single
request.

The patch adds a "bulk_process" boolean option, which if set results
in all rows to be submitted being submitted at once. Similar action
happens for verification.