Implement a pool of Swift clients.
Instatiating a Swift client is fairly expensive (~300ms on my laptop).
We should avoid this as much as possible and use a pool of
InternalClient objects instead.

This does require an API change where the ContainerCrawler library
will vend out the InternalClients for each of the handlers when
calling handle(). All of the consumers will need to be updated with
the new API.