Fix unicode name and exception handling.
When Swift objects or containers contain unicode characters, we need
to be careful when attempting to log them.

Secondly, the workers should always call task_done(), even if an
exception was raised. Catching the Exception class may not be
sufficient if an exception was derived from BaseException. Instead of
the specific exception class, we now catch anything that the "except"
clause will catch.