Remove deleted containers.
When a container is deleted, we should remove its CloudSync status
file. There is a window during which the container could be removed
and recreated, at which point we would not re-sync its content.

The commit contains a TODO, which states that we should keep track of
the container creation date and use it to reset the container status
(whenever it changes). This means that every call to sync will
require a HEAD request against the container. For the per-account
listing, we may be able to extract the creation date (X-Timestamp)
from the listing results (which would avoid additional HEAD
requests).