Add the ability for handlers to retry rows later.
There are times when an object cannot be processed at the moment. This
is not due to an error, but rather a scheduling or throttling
behavior. The patch adds a "RetryError" to express such intent.
ContainerCrawler does not advance the counters in this case, but also
does not log exceptions.