Support unicode characters in accounts/containers.
If there is a unicode character in the account or container name, we
should encode them as UTF-8 strings. Swift performs hash computations
to locate the container databases and the hashlib functions do not
handle unicode strings. The patch makes sure that the strings are
encoded before being passed on.