Add parallelism to discovering unprocessed rows.
Currently, container crawler serially queries every container for
new rows (objects). Before querying the next container database, it
waits for the already discovered objects to be processed. When there are
many containers with little work in each, this leads to poor
performance, as in the worst case it would be processing one object at a
time.

This patch introduces the concept of `enumerator_workers` -- green
threads that are used to discover new objects to be processed. This
allows the library to handle the case of numerous, but shallow
containers. The new configuration option `enumerator_workers` can be
used to set the number of these green threads (the default is 10).