Allow for verification of objects to be delayed.
Certain consumers may choose to allow for objects to be delayed in their
verification. The patch introduces an API change:
get_last_row() and save_last_row() are replaced with:
- get_last_processed_row()
- save_last_processed_row()
- get_last_verified_row()
- save_last_verified_row()

In addition, a new configuration option is added: "verification_slack".
This is a value in minutes of how long the verification may be delayed.