Add support for crawling sharded containers
For every container to be synced, check if it is sharded.
If so, then crawler needs to also crawl through
the hidden sharded containers.

The check currently requires a request to account server.
To prevent wasteful requests hitting the account servers
too many times, we limit listing of the hidden sharded
account once every 24 hours for those containers that
are not sharded. This time interval is configurable.