Use broker and container metadata to check shard status
Changed the code to check the shard status of containers to
lessen the amount of listing requests to the account servers.

Now the code will check for sharded containers in two different ways:
If the container database is local to the node, the container broker is used
to get the container metadata and thus the shard status. This does not
require any HTTP requests to Swift services.

There could be a scenario where a shard of a container might not be stored
on the same node as the root container, which would mean that on that
particular node we would never get the broker of the root container.
For this particular case, we use Swift's internal_client to get the
container metadata and check the shard status. If a container is sharded,
we will send an account listing request.