Allow for one-to-many mappings. (#37)
Change the tracking of in-progress containers to not be keyed only off
the account/container, but take into account the full settings
dictionary. This change does not do anything to validate that the
numerous mappings make sense (e.g. they could be in conflict with each
other or overwrite data if the discrepancy is in how the data is
handled). That problem is outside the ContainerCrawler purview.