Handle pre-2.18 Swift releases. (#38)
2.18 Swift introduced container sharding and the accompanying helper
methods. As the ContainerCrawler can be used with older versions of
Swift, we have to be careful about calling these methods.

Writing a test for this is a little bit difficult, as we need to have
some set of functional tests that we can execute against arbitrary
versions of Swift.